fix(labs): Moved panel header into separate component
