+++ modules/_labs/panel/react/index.ts
@@ -1,5 +1,5 @@
 import Panel from './lib/Panel';
-
+import PanelHeader from './lib/PanelHeader';
 export default Panel;
-export {Panel};
+export {Panel, PanelHeader};
 export * from './lib/Panel';
